Better prev/next logic when there are unshown images.
If the commandline includes only some of the images mentioned in an
existing Tags file, there will be invalid images (images that won't
be displayed) in the image list. If those end up at the end of the
list, then next_image to the end followed by prev_image wasn't
doing the right thing.